Rename math to mathEquations
